Adapted Sequential approch to now itterate over a hole (3x3) grid and apply each step once.
Changed calling of the individual function by looping over the grid and taking the grid[i][j] cell as the new middle cell.

When propagating a the propagation steps are written an a new propagation grid to avoid overwriting cells. Also added a check when the current cell is on a edge of the grid (currently the particals would not propagate furhter)